Some unsupported platform code has been removed
