+++ src/cmds/Makefile.am
@@ -240,9 +240,11 @@ qstop_CPPFLAGS = ${common_cflags}
 qstop_LDADD = ${common_libs}
 qstop_SOURCES = qstop.c ${common_sources}
 
-qsub_CPPFLAGS = ${common_cflags}
+qsub_CPPFLAGS = \
+	${common_cflags} \
+	-DX11_FEATURE
 qsub_LDADD = ${common_libs}
-qsub_SOURCES = qsub.c ${common_sources}
+qsub_SOURCES = qsub.c qsub_sup.c ${common_sources}
 
 qterm_CPPFLAGS = ${common_cflags}
 qterm_LDADD = ${common_libs}
